Manage the updated status.
For the script type,
If the livebox is not shown, try to merge the desc data.
And then send updated event to the provider when the provider prepare the buffer of a livebox.
Then we can prevent from lost of desc data.

For the PD.
If the PD is not opened, do not send updated event to the provider.
Just ignore it.

Change-Id: If346e84938fb5d59536f34ca7e3f308c6536ebca