Guarantees the destroy event of a livebox.
If there is only one loaded box in the slave,
The slave will be terminated when the last box is deleted.
But the box couldn't get the destroy event, if it is in updating status.

So while terminating the slave,
the process will forcely call the destroy & fainalize function.
Even if it is in updating status.

And then terminate the process.

So now the livebox can gets the destroy event correctly.

Change-Id: Ifd8544fb3be889b696c6ff87b8f60bd577b666fb